sprint 3: fix PHPStan level 9 errors in Repository (interfaces + cascade)
- Added @return EntityType[] to all RepositoryInterface methods (8 files)
- Fixed int|null cascade in EncuestaRepository (preguntaId, encuestaId)
- Fixed int|false cascade from strtotime in controllers (use (int) cast)
- Baseline: 375 → 335 errors

+++ src/Infrastructure/Web/Controller/DashboardController.php
@@ -44,7 +44,7 @@ class DashboardController extends BaseController
                 'id' => $d->getId(),
                 'titulo' => $d->getTitulo(),
                 'categoria' => $d->getCategoriaNombre() ?? '',
-                'subido' => $d->getCreatedAt() ? date('d/m/Y', strtotime($d->getCreatedAt())) : '',
+                'subido' => $d->getCreatedAt() ? date('d/m/Y', (int) strtotime($d->getCreatedAt())) : '',
             ], $recientes),
         ]);
     }
@@ -69,7 +69,7 @@ class DashboardController extends BaseController
                 'id' => $d->getId(),
                 'titulo' => $d->getTitulo(),
                 'categoria' => $d->getCategoriaNombre() ?? '',
-                'subido' => $d->getCreatedAt() ? date('d/m/Y', strtotime($d->getCreatedAt())) : '',
+                'subido' => $d->getCreatedAt() ? date('d/m/Y', (int) strtotime($d->getCreatedAt())) : '',
                 'activo' => $d->isActivo(),
             ], $recientes),
         ]);
